move tests and clarify names
